+++ src/config/records.php
@@ -5,4 +5,5 @@ return [
     'results_per_page' => 23,
     'admin_middleware'       => 'admin',
     'is_user_admin'     => 'isUserAdmin',
+    'use_rich_card'     => true,
 ];